feat: bookdeatilsview conectada con firebase

+++ app/src/main/java/com/example/libraryapp/ui/BookPreviews.kt
@@ -1,13 +1,11 @@
 package com.example.libraryapp.ui
 
-import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
 import androidx.compose.foundation.border
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
@@ -27,7 +25,6 @@ import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import androidx.navigation.compose.rememberNavController
 import coil.compose.AsyncImage
 import com.example.libraryapp.R
 import com.example.libraryapp.model.resources.Book
@@ -92,7 +89,7 @@ fun BookPreview (obra: Book){
             Row {
                 val rate = obra.score.toDouble()
                 val fillStars = Math.floor(rate / 2)
-                RatingBar (currentRating = fillStars.toInt())
+                RatingBar (currentRating = fillStars)
 
 /*                val rate = obra.score.toDouble()
                 val fillStars = Math.floor (rate / 2)
@@ -163,7 +160,7 @@ fun BookPreviewWide (obra : Book) {
                 Row {
                     val rate = obra.score.toDouble()
                     val fillStars = Math.floor(rate / 2)
-                    RatingBar (currentRating = fillStars.toInt())
+                    RatingBar (currentRating = fillStars)
 /*                    val halfStars = rate % 2
                     val unfilledStars = 5 - (fillStars + halfStars)
                     iconpainter(R.drawable.openmoji_star, fillStars.toInt())
